refactor: restructure monorepo with gRPC inter-service comms and Clean Architecture
- Move Maven wrappers to backend root; add per-service Dockerfiles
- api-gateway: remove local User/Role entities/repos, delegate auth to user-service via gRPC
- user-service: add gRPC server impl (UserGrpcServiceImpl) and JWT interceptor
- post-service: remove redundant User/Role entities, add post CRUD service/repo
- config-server: split into per-service property files (mail, post, user)
- Flutter hi_runner: scaffold Clean Architecture (domain/data/presentation, BLoC, get_it DI)
- Next.js: reorganize API layer (queries/mutations), move UI components

+++ backend/springboot-user-service/src/main/java/com/hiimtuan/user_service/service/UserServiceImpl.java
@@ -52,7 +52,7 @@ public class UserServiceImpl implements UserService {
     public UserResponseDto getProfile(String token) {
         String id = jwtService.extractUsername(token);
         
-        Optional<User> user = userRepository.findById(Integer.parseInt(id.trim()));
+        Optional<User> user = userRepository.findById(Long.parseLong(id.trim()));
 
         if (user.isEmpty()){
            throw new AccountException("Email doesn't exist");
